==> Code/02.feature_engineering.ipynb <==
# Feature Engineering
---
* Description:
>Feature engineering is the process of selecting, manipulating, and transforming raw data into features that can be used in supervised learning. 
In this notebook you can see our feature engineering pipeline. It depends on `feature_engineering_config.py` to lead the data flow.

---
## 1. Import Libraries

# basic
import pandas as pd
import numpy as np
import pickle

# outlier detection
from pyod.models.ecod import ECOD
from pyod.models.suod import SUOD
from pyod.models.lof import LOF
from pyod.models.iforest import IForest
from pyod.models.copod import COPOD
from sklearn.covariance import EllipticEnvelope

# feature selection
from xgboost import XGBRegressor

# homemade feature engineering
from feature_engineering import DealWithMissingValue, VariableTransformation, OutlierDetection,\
FeatureSelection, CreateGroupFeatureFromEachCol, CreateGroupFeatureFromAllCol, GeneratePolynomialFeatures, ReduceDimensionPCA, Standardization

# config
import feature_engineering_config as fe_config


pd.set_option('display.max_columns', 150)

## 2. Read xlsx File

train_data = pd.read_excel('../Data/2022-train-v2.xlsx')

## 3. Data Preparation

train_y = train_data[['sensor_point5_i_value', 'sensor_point6_i_value', 'sensor_point7_i_value', 'sensor_point8_i_value', 'sensor_point9_i_value', 'sensor_point10_i_value']]
train_x = train_data.drop(['sensor_point5_i_value', 'sensor_point6_i_value', 'sensor_point7_i_value', 'sensor_point8_i_value', 'sensor_point9_i_value', 'sensor_point10_i_value'], axis=1)
del train_data

train_x.head()

train_y.head()

## 4. Pipeline

train_process = fe_config.train_process
random_state = fe_config.random_state

# define nan
if fe_config.replace_zero_to_nan:
    train_x = train_x.replace({0: np.nan})

# missing values
if fe_config.deal_with_missing_value == 'drop_na':
    train_x = DealWithMissingValue(train_x).drop_na()
elif fe_config.deal_with_missing_value == 'imputation':
    train_x = DealWithMissingValue(train_x).imputation(strategy='most_frequent', train=train_process)
elif fe_config.deal_with_missing_value == 'k_neighbors_regressor':
    train_x = DealWithMissingValue(train_x).iterative_imputer(method='k_neighbors_regressor', train=train_process)
else:
    pass

# outlier detection on y
if fe_config.outlier_detection == 'ecod':
    temp = OutlierDetection(train_y).ecod(
        contamination=fe_config.ecod_contamination,
        threshold=fe_config.ecod_threshold,
        train=train_process
    )
elif fe_config.outlier_detection == 'suod':
    # initialized a group of outlier detectors for acceleration
    detector_list = [
        LOF(n_neighbors=15), LOF(n_neighbors=20), LOF(n_neighbors=25), LOF(n_neighbors=35),
        COPOD(),
        IForest(n_estimators=100), IForest(n_estimators=200)
    ]
    temp = OutlierDetection(train_y).suod(
        detector_list,
        threshold=fe_config.suod_threshold,
        train=train_process,
        verbose=False,
        n_jobs=4  # 如果是 mac 要設Ｎone
    )
elif fe_config.outlier_detection == 'elliptic_envelope':
    temp = OutlierDetection(train_y).elliptic_envelope(
        random_state=random_state,
        train=train_process
    )
else:
    pass

train_x['outlier_detection_from_y'] = temp
train_y['outlier_detection_from_y'] = temp
# remove outlier and drop outlier_detection_from_y
train_x = train_x.query("outlier_detection_from_y==False")
train_x = train_x.drop('outlier_detection_from_y', axis=1)
train_y = train_y.query("outlier_detection_from_y==False")
train_y = train_y.drop('outlier_detection_from_y', axis=1)

# variable transformation
large_skew_bool = train_x.skew().apply(abs) > fe_config.skew_threshold
with open(f'../Model/variable_transformation.pickle', 'wb' ) as f:
    pickle.dump(large_skew_bool, f)
for col in train_x._get_numeric_data().columns[large_skew_bool]:
    train_x[col] = VariableTransformation(train_x, col=col).transform(method=fe_config.variable_transformation, train=train_process)

    
    
    
    
# create group feature  
train_x['group'] = CreateGroupFeatureFromAllCol(train_x).kmeans_with_auto_k(
    standardization=fe_config.kmeans_standardization,
    k_range=fe_config.kmeans_k_range,
    random_state=random_state,
    parallel=True,
    parallel_verbose=0,
    train=train_process
).astype(str)


# convert each numeric data to category data
for col in train_x.columns:
    if col == 'group':
        pass
    else:
        train_x[col+'_group'] = CreateGroupFeatureFromEachCol(train_x[[col]]).kmeans_with_auto_k(
            standardization=fe_config.kmeans_standardization,
            k_range=fe_config.kmeans_k_range,
            random_state=random_state,
            parallel=True,
            parallel_verbose=0,
            train=train_process
        ).astype(str)
        
        
# generate polynomial and interaction features
if fe_config.generate_polynomial_features:
    train_x = GeneratePolynomialFeatures(train_x).get_dataframe(
        degree=fe_config.pf_degree,
        interaction_only=fe_config.pf_interaction_only,
        train=train_process
    )
    
if fe_config.reduce_dimension:
    train_x = ReduceDimensionPCA(train_x).pca(n_components=0.99, train=train_process)
    
    
if fe_config.standardization == 'standard_scaler':
    train_x = Standardization(train_x, prefix='x').standard_scaler(train=train_process)
elif fe_config.standardization == 'min_max_scaler':
    train_x = Standardization(train_x, prefix='x').min_max_scaler(train=train_process)


    
# Feature Selection
if 'variance' in fe_config.feature_selection:
    # removing features with zero variance
    train_x = FeatureSelection(train_x).variance(threshold=0, train=train_process)

# mutual information
if 'mutual_information' in fe_config.feature_selection:
    s = set()
    for col in train_x.columns:
        temp = FeatureSelection(test_x).mutual_information(k=50, train=train_process, y=train_y[col])
        s.update({temp.columns})
    test_x = test_x[list(s)]
    
# importance weight for feature selection
if 'importance_weight' in fe_config.feature_selection:
    estimator = XGBRegressor(n_jobs=-1)
    # 一定要用數結構的模型，因為這樣才能知道類別變數的重要性，而且分群變數(int)沒有做encoding，所以用tree based model 才不會有問題
    train_x = FeatureSelection(train_x).importance_weight(
        y=train_y,
        estimator=estimator,
        train=train_process
    )

# recursive feature elimination with cross-validation to select features
if 'recursive_feature_elimination' in fe_config.feature_selection:
    # 一定要用數結構的模型，因為這樣才能知道類別變數的重要性，而且分群變數(int)沒有做encoding，所以用tree based model 才不會有問題
    estimator = XGBRegressor(n_jobs=-1)
    train_x = FeatureSelection(train_x).recursive_feature_elimination(
        y=train_y,
        estimator=estimator,
        cv=fe_config.feature_selection_cv,
        min_features_to_select=fe_config.min_features_to_select,
        train=train_process,
        n_jobs=-1
    )

train_x.head()

with open(f'../Data/meta/train_x_after_feature_engineering.pickle', 'wb' ) as f:
    pickle.dump(train_x, f)

with open(f'../Data/meta/non_scaled_train_y_after_feature_engineering.pickle', 'wb' ) as f:
    pickle.dump(train_y, f)

if fe_config.standardization == 'standard_scaler'                     :
    train_y = Standardization(train_y, prefix='y').standard_scaler(train=fe_config.train_process)
elif fe_config.standardization == 'min_max_scaler':
    train_y = Standardization(train_y, prefix='y').min_max_scaler(train=fe_config.train_process)

with open(f'../Data/meta/train_y_after_feature_engineering.pickle', 'wb' ) as f:
    pickle.dump(train_y, f)
