==> Code/05.model_predict.ipynb <==
# Model Prediction
---
Description:

This notebook shows the test processing. In the end of content, you can get two predictive result by different method.
It depends on `model_predict_config.py` to lead the data flow.

---
## 1. Import Libraries

import pandas as pd
import numpy as np
import pickle

import model_predict_config as mp_config
import feature_engineering_config as fe_config
import model_chains_config

# feature engineering
from feature_engineering import DealWithMissingValue, VariableTransformation, OutlierDetection,\
FeatureSelection, CreateGroupFeatureFromEachCol, CreateGroupFeatureFromAllCol, GeneratePolynomialFeatures, ReduceDimensionPCA, Standardization

## 2. Read xlsx File

test_x = pd.read_excel('../Data/2022-test-v1.xlsx')

test_x.head()

test_x.shape

## 3. Pipeline

train_process = mp_config.train_process
random_state = mp_config.random_state

# define nan
if mp_config.replace_zero_to_nan:
    test_x = test_x.replace({0: np.nan})

# missing values
if mp_config.deal_with_missing_value == 'drop_na':
    test_x = DealWithMissingValue(test_x).drop_na()
elif mp_config.deal_with_missing_value == 'imputation':
    test_x = DealWithMissingValue(test_x).imputation(strategy='most_frequent', train=train_process)
elif mp_config.deal_with_missing_value == 'k_neighbors_regressor':
    test_x = DealWithMissingValue(test_x).iterative_imputer(method='k_neighbors_regressor', train=train_process)
else:
    pass


# variable transformation
with open(f'../Model/variable_transformation.pickle', 'rb') as f:
    large_skew_bool = pickle.load(f)

for col in test_x._get_numeric_data().columns[large_skew_bool]:
    test_x[col] = VariableTransformation(test_x, col=col).transform(method=mp_config.variable_transformation, train=train_process)

    
    
    
# create group feature  
test_x['group'] = CreateGroupFeatureFromAllCol(test_x).kmeans_with_auto_k(
    standardization=mp_config.kmeans_standardization,
    k_range=mp_config.kmeans_k_range,
    random_state=random_state,
    parallel=True,
    parallel_verbose=0,
    train=train_process
).astype(str)


# convert each numeric data to category data
for col in test_x.columns:
    if col == 'group':
        pass
    else:
        test_x[col+'_group'] = CreateGroupFeatureFromEachCol(test_x[[col]]).kmeans_with_auto_k(
            standardization=mp_config.kmeans_standardization,
            k_range=mp_config.kmeans_k_range,
            random_state=random_state,
            parallel=True,
            parallel_verbose=0,
            train=train_process
        ).astype(str)
        
        
# generate polynomial and interaction features
if mp_config.generate_polynomial_features:
    test_x = GeneratePolynomialFeatures(test_x).get_dataframe(
        degree=mp_config.pf_degree,
        interaction_only=mp_config.pf_interaction_only,
        train=train_process
    )
    
if mp_config.reduce_dimension:
    test_x = ReduceDimensionPCA(test_x).pca(n_components=0.99, train=train_process)
    
    
if mp_config.standardization == 'standard_scaler':
    test_x = Standardization(test_x, prefix='x').standard_scaler(train=train_process)
elif mp_config.standardization == 'min_max_scaler':
    test_x = Standardization(test_x, prefix='x').min_max_scaler(train=train_process)


    
# Feature Selection
test_x = test_x[mp_config.feature_select]

test_x.head()

## 4. Save test data after feature engineering

with open(f'../Data/meta/test_x_after_feature_engineering.pickle', 'wb' ) as f:
    pickle.dump(test_x, f)

## 5. Load model

* Muti-output Model

with open(f'../Model/catboost_all.pickle', 'rb') as f:
    muti_output_model = pickle.load(f)

* Single output Model

with open(f'../Model/catboost_single.pickle', 'rb' ) as f:
    single_output_model_dic = pickle.load(f)

* Stacking model

with open(f'../Model/stacking_model_dic.pickle', 'rb' ) as f:
    stacking_model_dic = pickle.load(f)

## 6. Prediction

with open(f'../Model/target_col_list.pickle', 'rb') as f:
    test_y_col = pickle.load(f)

* Muti-output Model

catboost_all_res = muti_output_model.predict(test_x)
catboost_all_res = pd.DataFrame(catboost_all_res, columns=test_y_col)

* Single output Model

array_li= [single_output_model_dic[col_y]['model'].predict(test_x).reshape(-1,1) for col_y in test_y_col]
catboost_single_res= np.concatenate(array_li, axis=1)
catboost_single_res = pd.DataFrame(catboost_single_res, columns=test_y_col)

* Stacking Model

data_dic = {}
for col in test_y_col:
    data_dic[col] = pd.concat([catboost_all_res[col], catboost_single_res[col]], axis=1)

array_li= [ stacking_model_dic[col].predict(data_dic[col]).reshape(-1,1) for col in test_y_col ]
all_res = np.concatenate(array_li, axis=1)

all_res = pd.DataFrame(all_res, columns=test_y_col)

# Load y-scaler model
with open(f'../Model/standard_scaler_y.pickle', 'rb') as f:
    scaler_y = pickle.load(f)

## 6. Load y Scaler Model

with open(f'../Model/standard_scaler_y.pickle', 'rb') as f:
    scaler_y = pickle.load(f)

all_res = scaler_y.inverse_transform(all_res)

final_df = pd.DataFrame(all_res, columns=test_y_col)

final_df.head()

final_df.shape

final_df.insert(0, 'No', list( range(1, 101, 1) ))

final_df.to_excel('../Output/111921_TestResult.xlsx', index=False)
