==> Code/96.model_chains(Albert).ipynb <==
# Executing multiple ML models in a chain
---
* Description:

>Model chains are defined by two or more models that execute sequentially, where the outputs of one model become the inputs to the next model down the chain.

* Direction of improvement:

>Build chain model by class.

---

## 1. Import Libraries

# basic
import model_chains_config
import pandas as pd
from sklearn.pipeline import Pipeline
import pickle

# preprocessing
from sklearn.preprocessing import StandardScaler
from sklearn.preprocessing import OneHotEncoder

# model
from sklearn.preprocessing import PolynomialFeatures
from sklearn.ensemble import AdaBoostRegressor
from sklearn.neighbors import KNeighborsRegressor
from catboost import CatBoostRegressor

%matplotlib inline
pd.set_option('display.float_format',lambda x: '%.2f' % x)

# import warnings
# warnings.filterwarnings('ignore')

def col_types(data):
    column_types = data.dtypes.to_frame().reset_index()
    column_types.columns = ['column_name', 'type']
    column_types.sort_values( by='type', inplace = True)
    return column_types

def missing_counts(data):
    missing = data.isnull().sum()
    missing = missing[missing >0]
    missing.sort_values(inplace=True)
    missing_count = pd.dataFrame({'column_name': missing.index, 'missing_count': missing.values})
    missing_count['percentage(%)'] = missing_Count['missing_count'].apply(lambda x:round(x/data.shape[0]*100,2))
    return missing_Count

with open('../Data/meta/train_x_after_feature_engineering.pickle', 'rb') as f:
    data_x = pickle.load(f)
data_x.head()

with open('../Data/meta/train_y_after_feature_engineering.pickle', 'rb') as f:
    data_y = pickle.load(f)
data_y.head()

# onehot encoding
for index , col in enumerate(data_x.columns.tolist()):
    if data_x[col].dtype == 'object':
        data_x = data_x.join(pd.get_dummies(data_x[col], prefix=col))
        data_x = data_x.drop(col, axis = 1)

display(col_types(data_x))

# lasso_params = {'sensor_point5_i_value': {'alpha': 500}, 
#                 'sensor_point6_i_value': {'alpha': 100}, 
#                 'sensor_point7_i_value': {'alpha': 500}, 
#                 'sensor_point8_i_value': {'alpha': 500}, 
#                 'sensor_point9_i_value': {'alpha': 500}, 
#                 'sensor_point10_i_value': {'alpha': 500}}

# ridge_params = {'sensor_point5_i_value': {'alpha': 0.1}, 
#                 'sensor_point6_i_value': {'alpha': 0.1}, 
#                 'sensor_point7_i_value': {'alpha': 0.1}, 
#                 'sensor_point8_i_value': {'alpha': 0.1}, 
#                 'sensor_point9_i_value': {'alpha': 1}, 
#                 'sensor_point10_i_value': {'alpha': 0.5}}

# xgb_params = {'sensor_point5_i_value': {'learning_rate': 0.2, 'max_depth': 3, 'min_child_weight': 10, 'n_estimators': 30, 'random_state': 1}, 
#               'sensor_point6_i_value': {'learning_rate': 0.2, 'max_depth': 5, 'min_child_weight': 10, 'n_estimators': 30, 'random_state': 1}, 
#               'sensor_point7_i_value': {'learning_rate': 0.2, 'max_depth': 3, 'min_child_weight': 10, 'n_estimators': 500, 'random_state': 1}, 
#               'sensor_point8_i_value': {'learning_rate': 0.2, 'max_depth': 3, 'min_child_weight': 3, 'n_estimators': 30, 'random_state': 1}, 
#               'sensor_point9_i_value': {'learning_rate': 0.2, 'max_depth': 5, 'min_child_weight': 10, 'n_estimators': 500, 'random_state': 1}, 
#               'sensor_point10_i_value': {'learning_rate': 0.2, 'max_depth': 20, 'min_child_weight': 10, 'n_estimators': 30, 'random_state': 1}}

kn_params = {
    'sensor_point5_i_value': {'leaf_size': 20, 'n_neighbors': 7, 'weights': 'distance'}, 
    'sensor_point6_i_value': {'leaf_size': 20, 'n_neighbors': 7, 'weights': 'distance'}, 
    'sensor_point7_i_value': {'leaf_size': 20, 'n_neighbors': 7, 'weights': 'distance'}, 
    'sensor_point8_i_value': {'leaf_size': 20, 'n_neighbors': 7, 'weights': 'distance'}, 
    'sensor_point9_i_value': {'leaf_size': 20, 'n_neighbors': 7, 'weights': 'distance'}, 
    'sensor_point10_i_value': {'leaf_size': 20, 'n_neighbors': 7, 'weights': 'distance'}
}

ada_params = {
    'sensor_point5_i_value': {'learning_rate': 0.2, 'n_estimators': 100, 'random_state': 1}, 
    'sensor_point6_i_value': {'learning_rate': 0.2, 'n_estimators': 100, 'random_state': 1}, 
    'sensor_point7_i_value': {'learning_rate': 0.5, 'n_estimators': 100, 'random_state': 1}, 
    'sensor_point8_i_value': {'learning_rate': 0.5, 'n_estimators': 300, 'random_state': 1}, 
    'sensor_point9_i_value': {'learning_rate': 0.2, 'n_estimators': 300, 'random_state': 1}, 
    'sensor_point10_i_value': {'learning_rate': 0.2, 'n_estimators': 100, 'random_state': 1}
}

cat_params = {
    'sensor_point5_i_value': {'depth': 8, 'eval_metric': 'RMSE', 'iterations': 500, 'l2_leaf_reg': 3, 'learning_rate': 0.03, 'random_state': 1}, 
    'sensor_point6_i_value': {'depth': 2, 'eval_metric': 'RMSE', 'iterations': 1000, 'l2_leaf_reg': 0.5, 'learning_rate': 0.03, 'random_state': 1}, 
    'sensor_point7_i_value': {'depth': 5, 'eval_metric': 'RMSE', 'iterations': 500, 'l2_leaf_reg': 3, 'learning_rate': 0.03, 'random_state': 1}, 
    'sensor_point8_i_value': {'depth': 5, 'eval_metric': 'RMSE', 'iterations': 500, 'l2_leaf_reg': 3, 'learning_rate': 0.03, 'random_state': 1}, 
    'sensor_point9_i_value': {'depth': 8, 'eval_metric': 'RMSE', 'iterations': 500, 'l2_leaf_reg': 0.5, 'learning_rate': 0.03, 'random_state': 1}, 
    'sensor_point10_i_value': {'depth': 8, 'eval_metric': 'RMSE', 'iterations': 500, 'l2_leaf_reg': 0.5, 'learning_rate': 0.03, 'random_state': 1}
}


order_list = [5, 3, 2, 1, 4, 0]

model_list = [
    Pipeline([
        ('poly',PolynomialFeatures(degree = 2)),
        ('std_scaler', StandardScaler()),
        ('Cat', CatBoostRegressor( **cat_params[model_chains_config.data_y_col[0]], verbose=False ))
    ]),          
    Pipeline([
        ('poly',PolynomialFeatures(degree = 2)),
        ('std_scaler', StandardScaler()),
        ('Cat', CatBoostRegressor( **cat_params[model_chains_config.data_y_col[1]], verbose=False ))
    ]),   
    Pipeline([
        ('poly',PolynomialFeatures(degree = 2)),
        ('std_scaler', StandardScaler()),
        ('KNR', KNeighborsRegressor( **kn_params[model_chains_config.data_y_col[2]] ))
    ]),
    Pipeline([
        ('poly',PolynomialFeatures(degree = 2)),
        ('std_scaler', StandardScaler()),
        ('KNR', KNeighborsRegressor(**kn_params[model_chains_config.data_y_col[3]] ))
    ]),
    Pipeline([
        ('poly',PolynomialFeatures(degree = 2)),
        ('std_scaler', StandardScaler()),
        ('Ada', AdaBoostRegressor(**ada_params[model_chains_config.data_y_col[4]] ))
    ]),
    Pipeline([
        ('poly',PolynomialFeatures(degree = 2)),
        ('std_scaler', StandardScaler()),
        ('Cat', CatBoostRegressor(**cat_params[model_chains_config.data_y_col[5]], verbose=False ))
    ])
]

data_y_col = ['sensor_point5_i_value', 'sensor_point6_i_value', 'sensor_point7_i_value',
              'sensor_point8_i_value', 'sensor_point9_i_value', 'sensor_point10_i_value']

x_pt = data_x.copy()
predict_y_train = pd.DataFrame({})
chain_model_dic = {}
for k in range(6):
    # 按照排列順序訓練模型
    chain_model_dic['model_y' + str(order_list[k])] = model_list[order_list[k]]
    chain_model_dic['model_y' + str(order_list[k])].fit(x_pt, data_y.iloc[:,order_list[k]])
             
    # 儲存預測結果
    predict_y_train[data_y_col[order_list[k]]] = chain_model_dic['model_y' + str(order_list[k])].predict(x_pt)
        
    # 把取得的預測值當作變數放進X
    x_pt[data_y_col[order_list[k]]] = chain_model_dic['model_y' + str(order_list[k])].predict(x_pt)

x_pt.shape

# 調整欄位順序
predict_y_train = predict_y_train[data_y_col]
predict_y_train

## Save model

with open('../model/chain_model_dic.pickle', 'wb') as f:
    pickle.dump(chain_model_dic, f)

# # 儲存預測結果(train)
# with open('../Output/predict_y_train(Albert).pickle', 'wb') as f:
#     pickle.dump(predict_y_train, f)

# # 儲存Model
# with open('../model/model_Y0(Albert).pickle', 'wb') as f:
#     pickle.dump(model_Y0, f)
    
# with open('../model/model_Y1(Albert).pickle', 'wb') as f:
#     pickle.dump(model_Y1, f)
    
# with open('../model/model_Y2(Albert).pickle', 'wb') as f:
#     pickle.dump(model_Y2, f)
    
# with open('../model/model_Y3(Albert).pickle', 'wb') as f:
#     pickle.dump(model_Y3, f)
    
# with open('../model/model_Y4(Albert).pickle', 'wb') as f:
#     pickle.dump(model_Y4, f)
    
# with open('../model/model_Y5(Albert).pickle', 'wb') as f:
#     pickle.dump(model_Y5, f)

## Predictions

# 迴圈依據套用模型計算預測值(Y)，並將預測結果放入X中做為新變數供下一輪模型預測時使用。(最佳模型順序[5, 3, 2, 1, 4, 0])

x_pt = data_x.copy()
predict_y_train = pd.DataFrame()

for k in range(6):
    model_order_li = order_list[k]
    col = data_y_col[model_order_li]
    
    predict_res = chain_model_dic['model_y'+ str(order_list[k])].predict(x_pt)
    predict_y_train[col] = predict_res
    # 新增 y 結果為 feature
    x_pt[col] = predict_res

# 調整欄位順序
predict_y_train = predict_y_train[data_y_col]
predict_y_train.head()

with open('../Data/meta/chain_model_res.pickle', 'wb') as f:
    pickle.dump(predict_y_train, f)

# # 迴圈依據套用模型計算預測值(Y)，並將預測結果放入X中做為新變數供下一輪模型預測時使用。(最佳模型順序[5, 3, 2, 1, 4, 0])

# x_pt = data_x.copy()
# predict_y_train = pd.DataFrame({})

# for k in range(6):
    
#     with open('../model/model_Y'+ str(model_chains_config.order_list[k]) + '(Albert).pickle', 'rb') as f:
#         model = pickle.load(f)
    
#     # 儲存預測結果
#     predict_y_train[model_chains_config.data_y_col[model_chains_config.order_list[k]]] = model.predict(x_pt)
#     # 把取得的預測值當作變數放進X
#     x_pt[model_chains_config.data_y_col[model_chains_config.order_list[k]]] = model.predict(x_pt)
    
#     # 清空 model
#     del model
    
# # 調整欄位順序
# predict_y_train = predict_y_train[model_chains_config.data_y_col]
# predict_y_train.head()

# Load y-scaler model
with open(f'../model/standard_scaler_y.pickle', 'rb') as f:
    scaler_y = pickle.load(f)
    
with open(f'../Data/meta/non_scaled_train_y_after_feature_engineering.pickle', 'rb') as f:
    non_scaled_train_y = pickle.load(f)

from utils.metric_utils import rmse_score

predictions = scaler_y.inverse_transform(predict_y_train)
targets = non_scaled_train_y
print(rmse_score(predictions, targets))
