==> Code/98.model_predict(Albert).ipynb <==
import model_chains_config
import ast
import pickle
import pandas as pd
import numpy as np

# import warnings
# warnings.filterwarnings('ignore')
pd.set_option('display.float_format',lambda x: '%.2f' % x)

def col_types(data):
    column_types = data.dtypes.to_frame().reset_index()
    column_types.columns = ['column_name', 'type']
    column_types.sort_values( by='type', inplace = True)
    return column_types

with open('../Data/meta/test_x_after_feature_engineering.pickle', 'rb') as f:
    data_x = pickle.load(f)
    
data_x

col_types(data_x)

# onehot encoding
for index , col in enumerate(data_x.columns.tolist()):
    if data_x[col].dtype == 'object':
        data_x = data_x.join(pd.get_dummies(data_x[col],prefix=col))
        data_x = data_x.drop(col, axis = 1)

# onehot encoding後有些train的欄位在測試集沒有，補上欄位

for i in model_chains_config.x_train_columns:
    if i not in data_x.columns.tolist():
        data_x[i] = np.zeros(100, dtype='uint8')

# 欄位排序
data_x = data_x[model_chains_config.x_train_columns]

col_types(data_x)

X_PT = data_x.copy()
predict_y_test = pd.DataFrame({})

for k in range(6):
    
    with open('../model/model_Y'+ str(model_chains_config.order_list[k]) + '(Albert).pickle', 'rb') as f:
        model = pickle.load(f)
    
    # 儲存預測結果
    predict_y_test[model_chains_config.data_y_col[model_chains_config.order_list[k]]] = model.predict(X_PT)
    # 把取得的預測值當作變數放進X
    X_PT[model_chains_config.data_y_col[model_chains_config.order_list[k]]] = model.predict(X_PT)
    
    # 清空 model
    del model
    
# 調整欄位順序
predict_y_test = predict_y_test[model_chains_config.data_y_col]
display(predict_y_test)

with open('../Output/predict_y_test(Albert).pickle', 'wb') as f:
    pickle.dump(predict_y_test, f)
