==> Code/99.single_model_stacking.ipynb <==
## Import Libraries

from sklearn.datasets import load_diabetes
from sklearn.linear_model import RidgeCV
from sklearn.svm import LinearSVR
from sklearn.ensemble import RandomForestRegressor
from xgboost import XGBRegressor
from sklearn.ensemble import StackingRegressor
from sklearn.linear_model import LinearRegression
from utils.metric_utils import rmse_score
import pandas as pd
import numpy as np
import pickle

## Load model

* Muti-output Model

with open(f'../Model/catboost_all.pickle', 'rb') as f:
    model_muti_output = pickle.load(f)

* Single output Model

with open(f'../Model/catboost_single.pickle', 'rb' ) as f:
    single_output_model_dic = pickle.load(f)

* Chain model

with open('../model/chain_model_dic.pickle', 'rb') as f:
    chain_model_dic = pickle.load(f)

## load Data

with open(f'../Data/meta/train_x_after_feature_engineering.pickle', 'rb') as f:
    train_x_after_fe = pickle.load(f)

with open(f'../Data/meta/train_y_after_feature_engineering.pickle', 'rb') as f:
    train_y_after_fe = pickle.load(f)

# You need to use original values of target variable to evaluate RMSE
with open(f'../Data/meta/non_scaled_train_y_after_feature_engineering.pickle', 'rb') as f:
    non_scaled_train_y = pickle.load(f)

train_x_after_fe.head(3)

train_y_after_fe.head(3)

non_scaled_train_y.head(3)

# RandomForestRegressor(n_estimators=10, random_state=42)

stack_model_dic = {}
for col in train_y_after_fe.columns:
    estimators = [
        (f'{col}_model_1', single_output_model_dic[col]['model'])
    ]
    reg = StackingRegressor(
        estimators=estimators,
        final_estimator=XGBRegressor(n_estimators=10, random_state=42),
        n_jobs=2
    )

    reg.fit(train_x_after_fe, train_y_after_fe[col])
    stack_model_dic[col] = {'model': reg}

stack_model_dic[col]['model']

with open(f'../Model/target_col_list.pickle', 'rb' ) as f:
    col_y_li = pickle.load(f)
    
array_li= [ stack_model_dic[col_y]['model'].predict(train_x_after_fe).reshape(-1,1) for col_y in col_y_li ]
all_res = np.concatenate(array_li, axis=1)

# Load y-scaler model
with open(f'../Model/standard_scaler_y.pickle', 'rb') as f:
    scaler_y = pickle.load(f)

predictions = scaler_y.inverse_transform(all_res)
targets = non_scaled_train_y
print(rmse_score(predictions, targets))
